+++ route/main_search.py
@@ -1,6 +1,4 @@
 from .tool.func import *
 
-def search_2(conn):
-    curs = conn.cursor()
-
+def main_search():
     return redirect('/search/' + url_pas(flask.request.form.get('search', 'test')))
