--- views/main_css/js/route/main_skin_set.js
@@ -1,9 +1,0 @@
-"use strict";
-
-function opennamu_do_main_skin_set() {
-
-}
-
-function opennamu_route_main_skin_set() {
-    
-}
